requirements.txt
blinkt
datetime
nre-darwin-py
